media-libs/mesa: don't check for create_memfd
The issue is that on older kernels the create_memfd system call
does not exist, however the new glibc does have the create_memfd
function. The mesa library calls the system call directly, though.
So, the check for create_memfd function during meson build config
succeeds (because it does not check for existance of syscall), but
the compilation fails. The workaround is to not check, which will
default to create_memfd not existing.