dev-python/firedrake: workaround for uGNI: no forking
Remove all usages of subprocess module, which calls fork().
The uGNI transport segfaults due to fork().

Remove usages in hdf5 (not needed so far).
Allow to not sniff compiler version by passing CCVER env var.

Part of the workaround is to run with a small input size (and with mpirun -n1),
to trigger the compilation, which will then be cached, and thus re-used for
runs with -n > 1, which must not call fork().